requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Storage
boto3==1.34.34
minio==7.2.3

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# AI/ML
openai==1.10.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

